Fix bugs

+++ Web_API/DataLayer/Models/Sensors/Speed.cs
@@ -2,8 +2,8 @@
 
 namespace DataLayer.Models.Sensors
 {
-    public class Speed : Sensor<int>
+    public class Speed : Sensor<float>
     {
-        public override int Value { get => base.Value; set => base.Value = value; }
+        public override float Value { get => base.Value; set => base.Value = value; }
     }
 }
